Se modifico los estilos de las cards del service
