+++ client/src/components/Servicie/SvgService.jsx
@@ -1,4 +1,3 @@
-import * as React from "react"
 
 const SvgComponent = (props) => (
   <svg
