+++ client/src/components/Servicie/styles/Card.module.css
@@ -1,7 +1,8 @@
 @import "../../VarGlobal.css";
+
 .card{
   background-color: var(--color-cuarto);
-  width: 100vh;
+  width: 28vh;
   height: 40vh;
   border-radius: 15px;
   margin-top: 3vh;
@@ -9,22 +10,38 @@
   -webkit-box-shadow: 8px 7px 22px 3px rgba(17,50,64,0.68);
   -moz-box-shadow: 8px 7px 22px 3px rgba(17,50,64,0.68);
 }
-
 .cardSec{
   background-color: var(--color-terciario);
-  width: 30vh;
+  width: 28vh;
   height: 40vh;
   border-radius: 15px;
   margin-top: 3vh;
   box-shadow: 8px 7px 22px 3px rgba(17,50,64,0.68);
--webkit-box-shadow: 8px 7px 22px 3px rgba(17,50,64,0.68);
--moz-box-shadow: 8px 7px 22px 3px rgba(17,50,64,0.68);
+  -webkit-box-shadow: 8px 7px 22px 3px rgba(17,50,64,0.68);
+  -moz-box-shadow: 8px 7px 22px 3px rgba(17,50,64,0.68);
+}
+.cardSec:hover{
+  cursor: pointer;
+  transition: 1.5s;
+  background-color: blur(4px);
+}
+.contentTitle{
+  margin: auto;
+  width: 90%;
+  
 }
 .title{
-  font-size: 20px;
+  color: var(--color-quinto);
+  font-size: 25px;
   text-align: end;
 }
-
+.contentParrafo{
+  width: 90%;
+  margin: auto;
+  height: auto;
+}
 .parrafo{
   text-align: start;
+  font-size: 20px;
+  color: var(--color-quinto);
 }
